Tweaked player fire rate, Tweaked enemy damage Set up UI to update based on local players health. Fixed new enemy prefab missing the player

+++ Assets/_GAME/_Scripts/Interface/IDamageable.cs
@@ -2,5 +2,5 @@ using UnityEngine;
 
 public interface IDamageable
 {   
-    public void ChangeHealth(int toChange, GameObject attacker);
+    public void ChangeHealth(float toChange, GameObject attacker);
 }
